* pcx.c: use size_t for memory buffer length (size_t is the appropiate type for this, unsigned int is not)  * change macro BUFFER_SIZE to static const JPEG_bufferSize (compiler will optimize this away)  * char* -> unsigned char*
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/trunk@1486 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/ivis_common/pcx.c
@@ -27,7 +27,7 @@
 
 #include "ivispatch.h"
 
-static const unsigned int PNG_BYTES_TO_CHECK = 4;
+static const size_t PNG_BYTES_TO_CHECK = 4;
 
 static void wzpng_read_data(png_structp ctx, png_bytep area, png_size_t size)
 {
